+++ Tests/InvariantSwiftMacroTests/Resources/Golden/PropertyTest/Basic.swift
@@ -1,4 +1,0 @@
-@PropertyTest("Basic test")
-func testBasicProperty(x: Int) {
-  x >= Int.min
-}